+++ app/lib/core/changelog.g.dart
@@ -11,6 +11,10 @@ const Changelog changelog = Changelog(<ChangelogEntry>[
   ChangelogEntry(
     version: 'Unreleased',
     sections: <ChangelogSection, List<String>>{
+      ChangelogSection.addedFeatures: <String>[
+        'Added Health Connect integration to synchronize weight measurements on Android',
+        'Added quick action to direclty launch app for adding a new weight',
+      ],
       ChangelogSection.otherChanges: <String>[
         'Introduce new M3E ButtonGroups Widget',
       ],